+++ tree-sitter-axiom/src/parser.c
@@ -48245,7 +48245,7 @@ TS_PUBLIC const TSLanguage *tree_sitter_axiom(void) {
     .metadata = {
       .major_version = 0,
       .minor_version = 3,
-      .patch_version = 1,
+      .patch_version = 2,
     },
   };
   return &language;